+++ CHANGELOG.md
@@ -8,6 +8,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
   - Webhooks: event creation in GraphQL when works are changed, event handling via Loco
 
+## [[0.13.9]](https://github.com/thoth-pub/thoth/releases/tag/v0.13.9) - 2025-04-10
+### Fixed
+  - [679](https://github.com/thoth-pub/thoth/issues/679) - Remove extraneous <custom_metadata> tag from Crossmark block in Crossref DOI deposit export when an Imprint has recorded a Crossmark DOI but a Work has no license or funding metadata.
+
+### Changed
+  - [683](https://github.com/thoth-pub/thoth/pull/683) - Upgrade rust to `1.86.0` in production `Dockerfile`
+
+### Security
+  - [683](https://github.com/thoth-pub/thoth/pull/683) - Upgrade `tokio` to v1.44.2
+  - [683](https://github.com/thoth-pub/thoth/pull/683) - Upgrade `openssl` to v0.10.72
+
+## [[0.13.8]](https://github.com/thoth-pub/thoth/releases/tag/v0.13.8) - 2025-03-26
+### Added
+  - [577](https://github.com/thoth-pub/thoth/pull/577) - Implement ONIX 3.1 "Thoth" specification (i.e. complete record reflecting all updates up to ONIX 3.1.2 that can be implemented based on existing data model)
+
+### Fixed
+  - [677](https://github.com/thoth-pub/thoth/pull/677) - Remove module directive in rapidocs import, which was cauisng CORS errors
+
 ## [[0.13.7]](https://github.com/thoth-pub/thoth/releases/tag/v0.13.7) - 2025-03-18
 ### Changed
   - [676](https://github.com/thoth-pub/thoth/pull/676) - Upgrade rust to `1.85.0` in production `Dockerfile`